+++ src/main/java/com/botmaker/sdk/authoring/TemplateLibrary.java
@@ -1,6 +1,5 @@
 package com.botmaker.sdk.authoring;
 
-import com.botmaker.plugin.api.authoring.ActivityModel;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import javax.imageio.ImageIO;